backend/requirements.txt
setuptools>=65.0.0
wheel>=0.40.0
Flask==2.3.3
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
# psycopg2-binary==2.9.9  # Removed for SQLite development (uncomment for PostgreSQL production)
python-dotenv==1.0.0
bcrypt==4.0.1
python-dateutil==2.8.2
requests==2.31.0
numpy>=1.26.0,<2.0.0
pandas>=2.1.0,<3.0.0
scikit-learn>=1.4.1
prophet==1.1.5
gunicorn==21.2.0